requirements.txt
google-cloud-logging==3.0.0
